Extract `replica_set/execute_read_operation.rs`
